requirements.txt
GitPython>=3.1.0
requests>=2.31.0
fastapi>=0.115.0
uvicorn>=0.30.0
celery>=5.3.0
redis>=5.0.0
pytest>=8.2.0
pytest-json-report>=1.5.0
pylint>=3.2.0
black>=24.8.0
ruff>=0.6.0
pydantic>=2.7.0
python-dotenv>=1.0.1
docker>=7.1.0
PyYAML>=6.0.1
typer>=0.12.5
jinja2>=3.1.4
markdown>=3.6

